backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
openai==1.57.4
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.3
tenacity==9.0.0
